refactor(organization): restructure create org drawer components
- Split monolithic CreateOrgDrawer component into modular files
- Extract form fields configuration into separate StepFields file
- Create dedicated FormStep component for better reusability
- Update imports in MainOrganization to use new structure
